windows: report 2

+++ Model/Models/Report/ReportOut.cpp
@@ -30,7 +30,7 @@ void ReportOut::setArticle(const qint64& article) {
 void ReportOut::setMysubject(const qint64& mysubject) {
     mysubject_ = mysubject;
 }
-void ReportOut::setCount(const qint64& count) {
+void ReportOut::setCount(const double& count) {
     count_ = count;
 }
 
@@ -45,7 +45,7 @@ const qint64& ReportOut::getArticle()const {
 const qint64& ReportOut::getMysubject()const {
     return mysubject_;
 }
-const qint64& ReportOut::getCount()const {
+const double& ReportOut::getCount()const {
     return count_;
 }
 
@@ -60,7 +60,7 @@ void ReportOut::read(const QJsonObject &jsonObj) {
     this->setCol(jsonObj[fld[ReportOut::Column::COL]].toInt());
     this->setArticle(jsonObj[fld[ReportOut::Column::ARTICLE]].toInt());
     this->setMysubject(jsonObj[fld[ReportOut::Column::MYSUBJECT]].toInt());
-    this->setCount(jsonObj[fld[ReportOut::Column::COUNT]].toInt());
+    this->setCount(jsonObj[fld[ReportOut::Column::COUNT]].toDouble());
 }
 ///Запись полей класса в JSON объкт.
